Backend/app/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
requests>=2.31
pandas>=2.0
numpy>=1.26
python-dateutil>=2.9
apscheduler>=3.10
sqlalchemy>=2.0
psycopg2-binary>=2.9
scikit-learn>=1.4
joblib>=1.3
